requirements.txt
python-telegram-bot==22.5
pony==0.7.19
APScheduler==3.11.2
